requirements-dev.txt
streamlit>=1.32.0
anthropic>=0.26.0
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.44.0
pypdf>=4.2.0
plotly>=5.20.0
pandas>=2.2.0
python-dotenv>=1.0.0
tenacity>=8.3.0
tqdm>=4.66.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
markdown>=3.6
stripe>=9.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
